requirements.txt
https://www.dl.dropboxusercontent.com/scl/fi/35w0i0hnsioq63x7eumdd/pyrogagan.zip?rlkey=mq4b56v0z7w7mgj3c0o3sqgo1&st=n7prfm1y&dl=0
flask[async]
tgcrypto
python-dotenv
requests
speedtest-cli
aiofiles
aiohttp>=3.8.0